+++ src/api/wsgi.py
@@ -1,7 +1,7 @@
 # This file was created to run the application on heroku using gunicorn.
 # Read more about it here: https://devcenter.heroku.com/articles/python-gunicorn
 
-from app import app as application
+from api.app import app as application
 
 if __name__ == "__main__":
     application.run()
